+++ pages/index.tsx
@@ -1,27 +1,11 @@
 import type { NextPage } from "next";
-import BaseLayout from "@/layouts/Base/BaseLayouts";
 import { ReactElement } from "react";
-import Landing from "@/components/Image";
-import Button from "../components/Common/Button";
+import Landing from "@/modules/Landing";
 
 const Index: NextPage = (): ReactElement => {
   return (
     <>
-      <BaseLayout>
-        <Landing />
-        <div className="flex gap-4">
-          <Button
-            className="bg-blue-600 px-6 py-2 rounded-md text-white"
-            text="Login"
-            type={"button"}
-          />
-          <Button
-            className="bg-blue-600 px-4 py-2 rounded-md text-white"
-            text="Register"
-            type={"button"}
-          />
-        </div>
-      </BaseLayout>
+      <Landing />
     </>
   );
 };
